aqara_api.py
import os
import requests
from dotenv import load_dotenv

# .env laden (falls noch nicht geladen)
load_dotenv()

AQARA_CLIENT_ID = os.environ.get("AQARA_CLIENT_ID")
AQARA_CLIENT_SECRET = os.environ.get("AQARA_CLIENT_SECRET")
AQARA_USERNAME = os.environ.get("AQARA_USERNAME")
AQARA_PASSWORD = os.environ.get("AQARA_PASSWORD")


class AqaraClient:
    """
    Minimaler Aqara-Client.
    Passe die URLs / Endpoints so an, wie du sie bisher in deiner Streamlit-Version hattest.
    Wichtig ist hier nur: KEINE Secrets im Code, alles kommt aus .env.
    """

    def __init__(self):
        if not all([AQARA_CLIENT_ID, AQARA_CLIENT_SECRET, AQARA_USERNAME, AQARA_PASSWORD]):
            raise ValueError("Aqara-Umgebungsvariablen fehlen – bitte .env prüfen.")

        self.client_id = AQARA_CLIENT_ID
        self.client_secret = AQARA_CLIENT_SECRET
        self.username = AQARA_USERNAME
        self.password = AQARA_PASSWORD

        self.access_token = None

    def authenticate(self):
        """
        Hier deine echte Aqara-Login-Logik einbauen.
        Das ist nur ein Platzhalter, damit die Struktur klar ist.
        """
        # Beispiel – bitte mit deinen echten Endpoints ersetzen:
        # resp = requests.post("https://api.aqara.com/auth", json={
        #     "client_id": self.client_id,
        #     "client_secret": self.client_secret,
        #     "username": self.username,
        #     "password": self.password,
        # })
        # resp.raise_for_status()
        # self.access_token = resp.json()["access_token"]
        pass

    def ensure_token(self):
        if not self.access_token:
            self.authenticate()

    def switch_on(self, device_id: str):
        """
        Beispiel-Methode zum Einschalten eines Gerätes.
        """
        self.ensure_token()
        # Hier deine echte Aqara-Request-Logik verwenden.
        # requests.post(..., headers={"Authorization": f"Bearer {self.access_token}"}, json=...)
        pass

    def switch_off(self, device_id: str):
        """
        Beispiel-Methode zum Ausschalten eines Gerätes.
        """
        self.ensure_token()
        # requests.post(...)
        pass